Reorganization for easier modification
