+++ unlock-app/src/__tests__/components/interface/keyChain/KeyDetails.test.tsx
@@ -1,11 +1,8 @@
 import React from 'react'
 import * as rtl from 'react-testing-library'
 import { MockedProvider } from '@apollo/react-testing'
-import {
-  Key,
-  OwnedKey,
-  KeyDetails,
-} from '../../../../components/interface/keyChain/KeyDetails'
+import { KeyDetails } from '../../../../components/interface/keyChain/KeyDetails'
+import { OwnedKey } from '../../../../components/interface/keyChain/KeychainTypes'
 import keyHolderQuery from '../../../../queries/keyHolder'
 
 const accountAddress = '0xAaAdEED4c0B861cB36f4cE006a9C90BA2E43fdc2'
@@ -21,59 +18,6 @@ const aKey: OwnedKey = {
     price: '50',
   },
 }
-/* const aSignature = {
- *   data: 'nice',
- *   signature: 'really nice',
- * } */
-
-describe('keyChain -- Key', () => {
-  it('should render the lock name', () => {
-    expect.assertions(0)
-    const signData = jest.fn()
-    const { getByText } = rtl.render(
-      <Key
-        signData={signData}
-        signature={null}
-        displayQR={jest.fn()}
-        accountAddress={accountAddress}
-        ownedKey={aKey}
-      />
-    )
-
-    getByText(aKey.lock.name)
-  })
-
-  it('should dispatch a payload to be signed', () => {
-    expect.assertions(1)
-    const signData = jest.fn()
-    const realDateNow = Date.now.bind(global.Date)
-    const dateNowStub = jest.fn(() => 1530518207007)
-    global.Date.now = dateNowStub
-    const { getByText } = rtl.render(
-      <Key
-        signData={signData}
-        signature={null}
-        displayQR={jest.fn()}
-        accountAddress={accountAddress}
-        ownedKey={aKey}
-      />
-    )
-
-    const button = getByText('Assert Ownership')
-    rtl.fireEvent.click(button)
-
-    expect(signData).toHaveBeenCalledWith(
-      JSON.stringify({
-        accountAddress,
-        lockAddress: aKey.lock.address,
-        timestamp: dateNowStub(),
-      }),
-      aKey.lock.address
-    )
-
-    global.Date.now = realDateNow
-  })
-})
 
 describe('keyChain -- KeyDetails', () => {
   it('should render loading state', () => {
