+++ unlock-app/src/components/interface/keyChain/KeyDetails.tsx
@@ -2,14 +2,11 @@ import React from 'react'
 import { useQuery } from '@apollo/react-hooks'
 import styled from 'styled-components'
 import keyHolderQuery from '../../../queries/keyHolder'
-import Media from '../../../theme/media'
 import 'cross-fetch/polyfill'
 import { DefaultError } from '../../creator/FatalError'
-import {
-  expirationAsDate,
-  durationsAsTextFromSeconds,
-} from '../../../utils/durations'
 import Loading from '../Loading'
+import { OwnedKey } from './KeychainTypes'
+import Key from './Key'
 
 export interface KeyDetailsProps {
   address: string
@@ -20,30 +17,6 @@ export interface KeyDetailsProps {
       data: string
       signature: string
     }
-  }
-}
-
-export interface OwnedKey {
-  id: string
-  expiration: string
-  keyId: string
-  lock: {
-    name: string
-    address: string
-    expirationDuration: string
-    tokenAddress: string
-    price: string
-  }
-}
-
-export interface KeyProps {
-  ownedKey: OwnedKey
-  accountAddress: string
-  signData: (data: any, id: any) => void
-  displayQR: (data: string) => void
-  signature: null | {
-    data: string
-    signature: string
   }
 }
 
@@ -105,141 +78,9 @@ export const NoKeys = () => {
   )
 }
 
-export class Key extends React.Component<KeyProps> {
-  constructor(props: KeyProps) {
-    super(props)
-  }
-
-  handleSignature = () => {
-    const {
-      accountAddress,
-      signData,
-      ownedKey: { lock },
-    } = this.props
-    const payload = JSON.stringify({
-      accountAddress,
-      lockAddress: lock.address,
-      timestamp: Date.now(),
-    })
-    signData(payload, lock.address)
-  }
-
-  showQRCode = () => {
-    const { displayQR, signature } = this.props
-    if (signature) {
-      displayQR(signature.data)
-    }
-  }
-
-  qrButton = () => {
-    const { signature } = this.props
-    if (signature) {
-      return (
-        <button type="button" onClick={this.showQRCode}>
-          Display QR Code
-        </button>
-      )
-    }
-    return (
-      <button type="button" onClick={this.handleSignature}>
-        Assert Ownership
-      </button>
-    )
-  }
-
-  render = () => {
-    const {
-      ownedKey: { lock, expiration },
-    } = this.props
-    return (
-      <Box>
-        <LockName>{lock.name}</LockName>
-        <LockExpirationDuration>
-          {durationsAsTextFromSeconds(parseInt(lock.expirationDuration))}
-        </LockExpirationDuration>
-        <ValidUntil>Valid Until</ValidUntil>
-        <KeyExpiration>{expirationAsDate(expiration)}</KeyExpiration>
-        {this.qrButton()}
-      </Box>
-    )
-  }
-}
-
-const Box = styled.div`
-  border: thin #dddddd solid;
-  width: 212px;
-  padding: 16px;
-  border-radius: 4px;
-  ${Media.phone`
-width: 100%;
-margin: 0 0 16px 0;
-`}
-  ${Media.nophone`
-width: 30%;
-margin: 0 16px 16px 0;
-`}
-  &:hover {
-    border: thin #aaaaaa solid;
-    box-shadow: 0px 0px 10px 3px rgba(221, 221, 221, 1);
-  }
-`
-
 const Container = styled.div`
   display: flex;
   flex-direction: row;
   flex-wrap: wrap;
   max-width: 100%;
-`
-
-const LockName = styled.div`
-  font-family: IBM Plex Sans;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 15px;
-  line-height: 19px;
-  /* or 127% */
-
-  display: flex;
-  align-items: center;
-  color: #4d8be8;
-`
-
-const LockExpirationDuration = styled.div`
-  font-family: IBM Plex Sans;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 15px;
-  line-height: 19px;
-  /* identical to box height, or 127% */
-
-  display: flex;
-  align-items: center;
-
-  /* Grey 4 */
-
-  color: #333333;
-  margin-top: 8px;
-`
-
-const ValidUntil = styled.div`
-  font-family: IBM Plex Sans;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 8px;
-  line-height: 10px;
-  /* identical to box height */
-
-  letter-spacing: 1px;
-  text-transform: uppercase;
-  color: #a6a6a6;
-  margin-top: 8px;
-`
-
-const KeyExpiration = styled.div`
-  font-family: IBM Plex Sans;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 16px;
-  line-height: 20px;
-  color: #333333;
 `
