+++ unlock-app/src/components/interface/keyChain/KeychainTypes.ts
@@ -1,0 +1,12 @@
+export interface OwnedKey {
+  id: string
+  expiration: string
+  keyId: string
+  lock: {
+    name: string
+    address: string
+    expirationDuration: string
+    tokenAddress: string
+    price: string
+  }
+}
